Fix pink materials, font crash, input null ref, bootstrapper timing
FIXES:
- Pink materials: Shader.Find("Universal Render Pipeline/Lit")
  returns null in some Unity 6 configs. Added EditorShaderHelper
  with cascading fallback: URP Lit → Simple Lit → Standard → Unlit.
  Uses _BaseColor for URP, .color for Standard. All 17 Shader.Find
  calls across 4 editor scripts replaced.
- Font crash: "Arial.ttf is no longer a valid built in font" in
  Unity 6. CombatHUD.CreateText now uses GetSafeFont() with try/catch
  fallback chain: LegacyRuntime → Arial → OS font → any OS font.
- PlayerInputHandler NullRef: _playerInput.actions was null when no
  InputActionAsset assigned. Added null check before FindAction.
- DealingBootstrapper: increased delay to 0.5s and added 5-retry
  loop to handle late player initialization.
- AddClout cast fix: float → int for ReputationManager calls
- SupplierNPC: added missing WantedSystem using

GENERATED ASSETS (from Unity editor run):
- 5 Product SOs: Weed, Crystal, Powder, Pills, Shroom
- 2 Supplier SOs: Lil D, Mr Kim
- Updated TestArena.unity scene with dealing NPCs

+++ Assets/_Project/Scripts/Empire/Dealing/DealManager.cs
@@ -183,10 +183,10 @@ namespace Clout.Empire.Dealing
                 float clout = totalCash >= 500f
                     ? ReputationManager.CloutValues.BigDeal
                     : ReputationManager.CloutValues.CompleteDeal;
-                rep.AddClout(clout, "drug deal");
+                rep.AddClout((int)clout, "drug deal");
 
                 if (rep.cloutRank.Value == 0) // First deal ever
-                    rep.AddClout(ReputationManager.CloutValues.FirstSale, "first sale");
+                    rep.AddClout((int)ReputationManager.CloutValues.FirstSale, "first sale");
             }
 
             // Heat
